flask_redis_editor/pyproject.toml
[tool.poetry]
name = "flask-redis-editor"
version = "0.1.0"
description = ""
authors = ["Viacheslav <stepvg@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
redis = "^5.0.3"
gevent = "^24.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
